chore: evm e2e with new submitter
