+++ rust/main/utils/run-locally/src/main.rs
@@ -29,7 +29,7 @@ use std::{
 };
 
 use ethers_contract::MULTICALL_ADDRESS;
-use hyperlane_core::{PendingOperationStatus, ReorgEvent, ReprepareReason};
+use hyperlane_core::{PendingOperationStatus, ReorgEvent, ReprepareReason, SubmitterType};
 use logging::log;
 pub use metrics::fetch_metric;
 use once_cell::sync::Lazy;
@@ -96,6 +96,8 @@ const FAILED_MESSAGE_COUNT: u32 = 1;
 
 const RELAYER_METRICS_PORT: &str = "9092";
 const SCRAPER_METRICS_PORT: &str = "9093";
+
+pub const SUBMITTER_TYPE: SubmitterType = SubmitterType::Lander;
 
 type DynPath = Box<dyn AsRef<Path>>;
 
@@ -377,7 +379,7 @@ fn main() -> ExitCode {
     let mut test_passed = wait_for_condition(
         &config,
         loop_start,
-        || termination_invariants_met(&config, starting_relayer_balance),
+        || termination_invariants_met(&config, starting_relayer_balance, SUBMITTER_TYPE),
         || !SHUTDOWN.load(Ordering::Relaxed),
         || long_running_processes_exited_check(&mut state),
     );
@@ -482,6 +484,9 @@ fn create_relayer(rocks_db_dir: &TempDir) -> Program {
         .hyp_env("DB", relayer_db.to_str().unwrap())
         .hyp_env("CHAINS_TEST1_SIGNER_KEY", RELAYER_KEYS[0])
         .hyp_env("CHAINS_TEST2_SIGNER_KEY", RELAYER_KEYS[1])
+        .hyp_env("CHAINS_TEST1_SUBMITTER", SUBMITTER_TYPE.to_string())
+        .hyp_env("CHAINS_TEST2_SUBMITTER", SUBMITTER_TYPE.to_string())
+        .hyp_env("CHAINS_TEST3_SUBMITTER", SUBMITTER_TYPE.to_string())
         .hyp_env("RELAYCHAINS", "invalidchain,otherinvalid")
         .hyp_env("ALLOWLOCALCHECKPOINTSYNCERS", "true")
         .hyp_env(
